Fix typos and improve clarity of FAQ copy (#71)
I initially forked this to fix a typo on this page, but I couldn't resist taking a crack at
doing a full edit. The changes I made are intended to improve the clarity of the page overall,
and are mostly focused on areas that I found the phrasing to be awkward, vague, or wordy. I
kept as much of the original copy where possible and tried to maintain the approachable tone.
My hope is that the improved clarity might help increase accessibility, especially for those
whose first language is not English.